add yolov6, skip opencv preprocess bug caused by bmopencv in cv::resize when use ROI to cv::Mat
Reviewed-on: https://gerrit-ai.sophgo.vip:8443/136455
Tested-by: 福全 柯 <fuquan.ke@sophgo.com>

+++ modules/framework/net/bm168x_net.cpp
@@ -239,7 +239,7 @@ int32_t BM168xNet::addInput(const std::string &name) {
       name.c_str(), element_bytes, shape[0], shape[1], shape[2], shape[3],
       input_output_tensor_infos_[name].qscale,
       input_output_tensor_infos_[name].zero_point,
-      input_output_tensor_infos_[name].data_type);
+      static_cast<int>(input_output_tensor_infos_[name].data_type));
   return 0;
 }
 
@@ -266,7 +266,7 @@ int32_t BM168xNet::addOutput(const std::string &name) {
       name.c_str(), element_bytes, shape[0], shape[1], shape[2], shape[3],
       input_output_tensor_infos_[name].qscale,
       input_output_tensor_infos_[name].zero_point,
-      input_output_tensor_infos_[name].data_type);
+      static_cast<int>(input_output_tensor_infos_[name].data_type));
   return 0;
 }
